move sio_update_game
